FIX order of draw calls in SankeyController
Canvas paints new draw operations on top of exisiting ones. Due to the fact that the _drawLabels method is invoked first, the text must compete with the transparent Flows drawn over it which limits the range of colors that are legible/visible (as well as fidelity of the text) within the chart. I changed the order of the calls such that _drawLabels occurs after the for expression for drawing the Flows